added unit-testing

+++ Source/MonitoringService/MonitoringService.Contracts/StatisticsResponse.cs
@@ -8,7 +8,7 @@ public class StatisticsResponse
     /// <summary>
     /// Уникальный идентификатор
     /// </summary>
-    public ulong Id { get; set; }
+    public int Id { get; set; }
     
     /// <summary>
     /// Имя устройства
